# api/openapi.yaml
openapi: 3.1.0
info:
  title: jtpost API
  version: 0.10.0
  description: |
    REST API для jtpost — CLI/web-сервиса управления постами для Telegram.

    Текущая стабильная версия — `/api/v1/`. Legacy `/api/` сохранён для backward-compat.

    Authentication:
    - `bearerAuth` — Personal Access Token (`jtpat_<8>_<24>`) в `Authorization: Bearer ...`.
    - `cookieAuth` — Session cookie `jtpost_session` (HttpOnly, Secure, SameSite=Lax).
  license:
    name: MIT
servers:
  - url: /api/v1
    description: Versioned API (stable)
  - url: /api
    description: Legacy (без версии, backward-compat)
...
paths:
...
  /outbox:
    get:
      summary: List outbox entries
      operationId: listOutbox
      tags: [outbox]
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, in_flight, done, failed]
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 500
      responses:
        '200':
          description: Outbox entries
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/OutboxFullEntry'}
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          description: Outbox not configured
          content:
            text/plain:
              schema: {type: string}
...
  /outbox/{id}/retry:
    parameters:
      - name: id
        in: path
        required: true
        schema: {type: string, format: uuid}
    post:
      summary: Force retry outbox entry
      description: Сбрасывает entry в pending, attempts=0, next_attempt_at=now.
      operationId: retryOutbox
      tags: [outbox]
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        '202':
          description: Retry scheduled
          content:
            application/json:
              schema: {$ref: '#/components/schemas/OutboxFullEntry'}
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Entry already done
          content:
            text/plain:
              schema: {type: string}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: PAT
      description: Personal Access Token формата `jtpat_<8>_<24>`
    cookieAuth:
      type: apiKey
      in: cookie
      name: jtpost_session
...
  responses:
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    OutboxFullEntry:
      type: object
      required: [id, post_id, tenant_id, kind, status, attempts, max_attempts, next_attempt_at, created_at, updated_at]
      properties:
        id: {type: string, format: uuid}
        post_id: {type: string, format: uuid}
        tenant_id: {type: string, format: uuid}
        kind: {type: string}
        status: {type: string, enum: [pending, in_flight, done, failed]}
        attempts: {type: integer}
        max_attempts: {type: integer}
        next_attempt_at: {type: string, format: date-time}
        last_error: {type: string}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        details:
          type: string